Order, Instant, ISO 8601 Part-1

+++ src/main/java/com/example/course/resources/UserResource.java
@@ -25,7 +25,7 @@ public class UserResource {
     }
 
     @GetMapping(value = "/{id}")
-    public ResponseEntity<User> findById(@PathVariable Long id) {
+    public ResponseEntity<User> findById(@PathVariable Integer id) {
         User obj = service.findById(id);
         return ResponseEntity.ok().body(obj);
